MRP - TCO optimization
Create MRP - TCO optimization
In this case, we're tasked with the mission to propose the optimal amount of cars to produce each month to minimize TCO (total cost of ownership), which is the total of purchasing, holding and production costs. Inputs of next year's forecasted demand, unit holding costs and production costs (computed as total costs of materials needed to build a car) is given.